+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 
 ## Unreleased - 2021-??-??
 
+### Changed
+* Bump ASM from 9.0 to 9.1 supporting JDK17
+
 ## 4.2.1 - 2021-02-04
 
 ### Fixed